add grpc for UserTradeService

+++ backend/src/main/java/com/example/backend/dto/AddStrategyRequest.java
@@ -4,6 +4,6 @@ import com.example.backend.model.Strategy;
 
 public record AddStrategyRequest (
     long tokenId,
-    double amount,
+    Integer amount,
     Strategy.StrategyType type
 ){}
